Move PIS to dedicated module
